+++ src/Features/SupportsCreatedAt.php
@@ -5,5 +5,6 @@ namespace Parable\Orm\Features;
 interface SupportsCreatedAt
 {
     public function setCreatedAt(string $createdAt): void;
+
     public function markCreatedAt(): void;
 }